+++ packages/x-components/src/components/layouts/multi-column-max-width-layout.vue
@@ -104,6 +104,8 @@
   /**
    * Component for use as Layout to be filled with the rest of the Components.
    *
+   * @deprecated - The layout has been deprecated in favor of using new XDS layout.
+   *
    * @public
    */
   @Component({
@@ -157,7 +159,7 @@
 </style>
 
 <style lang="scss" scoped>
-  @import '../../design-system/utilities/dev-mode';
+  @import '../../design-system-deprecated/utilities/dev-mode';
 
   .x-layout {
     // custom properties