+++ packages/x-components/src/components/layouts/single-column-layout.vue
@@ -87,6 +87,8 @@
   /**
    * Component for use as Layout to be filled with the rest of the Components.
    *
+   * @deprecated - The layout has been deprecated in favor of using new XDS layout.
+   *
    * @public
    */
   @Component({
@@ -104,7 +106,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '../../design-system/utilities/dev-mode';
+  @import '../../design-system-deprecated/utilities/dev-mode';
 
   .x-layout {
     display: grid;