+++ packages/x-components/src/views/home/aside.vue
@@ -1,7 +1,5 @@
 <template>
-  <div
-    class="x-list x-list--padding-05 x-list--padding-top x-list--gap-06 x-list--border x-list--border-top"
-  >
+  <div class="x-flex x-flex-col x-gap-24">
     <FacetsProvider :facets="staticFacets" />
     <ClearFilters />
     <SelectedFiltersList>
@@ -11,10 +9,10 @@
     </SelectedFiltersList>
 
     <!-- Facets -->
-    <Facets class="x-list--gap-06">
+    <Facets class="x-gap-24">
       <!--  Hierarchical Facet    -->
       <template #hierarchical-facet="{ facet }">
-        <BaseHeaderTogglePanel class="x-facet">
+        <BaseHeaderTogglePanel headerClass="x-w-full x-flex x-justify-between x-py-8">
           <template #header-content>
             <span class="x-truncate">{{ facet.label }}</span>
             <ChevronDown />
@@ -25,7 +23,7 @@
               <HierarchicalFilter
                 :filter="filter"
                 :data-test="`${facet.label}-filter`"
-                childrenFiltersClass="x-margin--left-05"
+                childrenFiltersClass="x-ml-16"
               />
             </FiltersList>
           </SlicedFilters>
@@ -34,7 +32,7 @@
 
       <!--  Range Facet    -->
       <template #number-range-facet="{ facet }">
-        <BaseHeaderTogglePanel class="x-facet">
+        <BaseHeaderTogglePanel headerClass="x-w-full x-flex x-justify-between x-py-8">
           <template #header-content>
             <span :data-test="facet.label" class="x-truncate">{{ facet.label }}</span>
             <ChevronDown />
@@ -71,11 +69,11 @@
 
       <!--  Default Facet    -->
       <template #default="{ facet }">
-        <BaseHeaderTogglePanel class="x-facet">
+        <BaseHeaderTogglePanel headerClass="x-w-full x-flex x-py-8 x-gap-8">
           <template #header-content>
             <span :data-test="facet.label" class="x-truncate">{{ facet.label }}</span>
             <span data-test="total-filters">{{ facet.filters.length }}</span>
-            <ChevronDown />
+            <ChevronDown class="x-ml-auto" />
           </template>
 
           <!-- Filters -->