+++ packages/x-components/src/x-modules/history-queries/components/my-history.vue
@@ -1,5 +1,5 @@
 <template>
-  <component :is="animation" v-if="hasHistoryQueries" class="x-my-history x-list" tag="ul">
+  <component :is="animation" v-if="hasHistoryQueries" class="x-my-history" tag="ul">
     <li
       v-for="(historyQueries, date) in groupByDate"
       :key="date"
@@ -37,10 +37,7 @@
                   @binding {() => string} formatTime - Callback to format time to `hh:mm [PM/AM]`
             -->
                 <slot name="suggestion-content" v-bind="{ suggestion, index, formatTime }">
-                  <div class="x-list x-list--vertical">
-                    <span>{{ suggestion.query }}</span>
-                    <span>{{ formatTime(suggestion.timestamp) }}</span>
-                  </div>
+                  {{ suggestion.query }} - {{ formatTime(suggestion.timestamp) }}
                 </slot>
               </template>
               <template #remove-button-content>
@@ -196,6 +193,13 @@
     }
   }
 </script>
+
+<style lang="scss" scoped>
+  .x-my-history {
+    display: flex;
+    flex-flow: column nowrap;
+  }
+</style>
 
 <docs lang="mdx">
 ## Events