+++ packages/x-components/vue.config.js
@@ -1,11 +1,6 @@
-const glob = require('glob');
-
 module.exports = {
   configureWebpack: {
-    entry: [
-      ...glob.sync('src/design-system/**/*.scss').map(style => `./${style}`),
-      './src/main.ts'
-    ],
+    entry: ['./src/main.ts'],
     output: {
       devtoolModuleFilenameTemplate: ({ resourcePath, hash, query }) =>
         resourcePath.match(/\.vue$/) &&